perf(v0.18 WS-E): storage packing + SLOAD caching gas optimizations (#82/#81/#80/#79)
#82 — AAStarGlobalGuard.TokenConfig: 3 slots → 2
  Pack tier1Limit+tier2Limit as uint128 (one slot); dailyLimit stays uint256.
  Explicit TierLimitTooLarge bound check in _validateTokenConfig prevents silent
  truncation from any wider source. Saves 1 cold SSTORE (~20k) on fresh config writes
  and 1 SLOAD (~2k) per recordTokenSpend tier read.

#81 — executeBatch guard-address SLOAD caching
  Read address(guard) ONCE per execute/executeBatch/executeFromExecutor and thread it
  into _enforceGuard/_checkTokenGuard instead of re-SLOADing slot 4 per batched call.
  Behavior identical (guard immutable post-init).

#80 — SessionKeyValidator.SessionState: 3 slots → 1
  Pack windowStart(uint48)+callCount(uint32)+prevCount(uint32) into one slot. WS-C
  weighted sliding-window math unchanged (operates in uint256; narrow fields widen on
  read, narrow back via bounded casts). Saves up to ~44k gas/velocity record (3→1 SSTORE).

#79 — _popcount / _countBits already SWAR assembly; added PopcountAssembly.t.sol proving
  assembly == reference clear-lowest-bit loop (exhaustive low range + all single bits +
  guardian-bitmap domain + 256-run fuzz; all-ones edge asserted).

Factory size: TokenConfig uint128 masking embedded 3× pushed AAStarAirAccountFactoryV7
  initcode over EIP-3860; deduplicated the three new AAStarGlobalGuard(...) sites into
  _deployGuard() → initcode back under limit (49,134 < 49,152).

Storage: forge inspect confirms NO top-level slot shift on account/guard/validator/forceexit.
Tests: 798 cancun (794 baseline + 4 new popcount), 0 failed; 22 prague, 0 failed.
ABI docs + full-ABI regenerated (getter return-type narrowing + new error selector).

+++ test/AAStarGlobalGuardM5.t.sol
@@ -28,7 +28,9 @@ contract AAStarGlobalGuardM5Test is Test {
     bytes4 constant TSFR_FROM = 0x23b872dd;
 
     // Token decimals
-    uint256 constant USDC_DEC = 1e6;
+    // #82: uint128 so `N * USDC_DEC` matches the packed uint128 TokenConfig tier fields.
+    //      Other uses (recordTokenSpend amount, assertEq) widen to uint256 implicitly.
+    uint128 constant USDC_DEC = 1e6;
     uint256 constant ETH_DEC  = 1 ether;
 
     // ─── Test Setup ───────────────────────────────────────────────────
